Updated Trade model quantity field type

+++ backtest/models.py
@@ -56,7 +56,7 @@ class Trade(models.Model):
     leg_id = models.CharField(max_length=100)    
     timestamp = models.DateTimeField()
     symbol = models.CharField(max_length=50)     
-    quantity = models.IntegerField()
+    quantity = models.DecimalField(max_digits=10, decimal_places=4)
     price = models.DecimalField(max_digits=10, decimal_places=4)
     cost = models.DecimalField(max_digits=10, decimal_places=4)
     action = models.CharField(max_length=10)  # Assuming 'direction' is a string like 'buy' or 'sell'
